+++ src/main/java/com/abdulmajid/expensetracker/dto/LoanResponse.java
@@ -2,21 +2,22 @@ package com.abdulmajid.expensetracker.dto;
 
 import com.abdulmajid.expensetracker.model.User;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 public class LoanResponse {
     private Integer id;
-    private Integer amount;
+    private BigDecimal amount;
     private String lender;  // Who provided the loan
     private String borrower; // Optional, if applicable
-    private Double interestRate;
+    private BigDecimal interestRate;
     private String loanType;
     private Integer tenureMonths;
     private Date startDate;
     private Date dueDate;
     private String paymentMode;
     private String isPaid;
-    private Double remainingBalance;
+    private BigDecimal remainingBalance;
     private String status;
     private String note;
     private String day;
@@ -28,9 +29,9 @@ public class LoanResponse {
     public LoanResponse() {
     }
 
-    public LoanResponse(Integer id, Integer amount, String lender, String borrower, Double interestRate,
+    public LoanResponse(Integer id, BigDecimal amount, String lender, String borrower, BigDecimal interestRate,
                         String loanType, Integer tenureMonths, Date startDate, Date dueDate,
-                        String paymentMode, String isPaid, Double remainingBalance, String status,
+                        String paymentMode, String isPaid, BigDecimal remainingBalance, String status,
                         String note, String day, Date date, Date createdAt, Date updatedAt, User user) {
         this.id = id;
         this.amount = amount;
@@ -61,11 +62,11 @@ public class LoanResponse {
         this.id = id;
     }
 
-    public Integer getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
-    public void setAmount(Integer amount) {
+    public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
 
@@ -85,11 +86,11 @@ public class LoanResponse {
         this.borrower = borrower;
     }
 
-    public Double getInterestRate() {
+    public BigDecimal getInterestRate() {
         return interestRate;
     }
 
-    public void setInterestRate(Double interestRate) {
+    public void setInterestRate(BigDecimal interestRate) {
         this.interestRate = interestRate;
     }
 
@@ -141,11 +142,11 @@ public class LoanResponse {
         this.isPaid = isPaid;
     }
 
-    public Double getRemainingBalance() {
+    public BigDecimal getRemainingBalance() {
         return remainingBalance;
     }
 
-    public void setRemainingBalance(Double remainingBalance) {
+    public void setRemainingBalance(BigDecimal remainingBalance) {
         this.remainingBalance = remainingBalance;
     }
